# refactor(types/parser): parse by fragment and process in pipeline