--- a/README.adoc
+++ b/README.adoc
@@ -76,11 +76,11 @@ Libasciidoc provides 2 functions to convert an Asciidoc content into HTML:
 
 1. Converting an `io.Reader` into an HTML document:
 
-    Convert(r io.Reader, output io.Writer, config configuration.Configuration) (types.Metadata, error) 
+    Convert(r io.Reader, output io.Writer, config *configuration.Configuration) (types.Metadata, error) 
 
 2. Converting a file (given its name) into an HTML document:
 
-   ConvertFile(output io.Writer, config configuration.Configuration) (types.Metadata, error)
+   ConvertFile(output io.Writer, config *configuration.Configuration) (types.Metadata, error)
 
 where the returned `types.Metadata` object contains the document's title which is not part of the generated HTML `<body>` part, as well as the other attributes of the source document.