--- a/cmd/libasciidoc/root_cmd.go
+++ b/cmd/libasciidoc/root_cmd.go
@@ -119,8 +119,8 @@ func getOut(cmd *cobra.Command, sourcePath, outputName string) (io.Writer, close
 }
 
 // converts the `name`, `!name` and `name=value` into a map
-func parseAttributes(attributes []string) map[string]string {
-	result := make(map[string]string, len(attributes))
+func parseAttributes(attributes []string) map[string]interface{} {
+	result := make(map[string]interface{}, len(attributes))
 	for _, attr := range attributes {
 		data := strings.Split(attr, "=")
 		if len(data) > 1 {