--- a/cmd/libasciidoc/root_cmd.go
+++ b/cmd/libasciidoc/root_cmd.go
@@ -53,7 +53,6 @@ func NewRootCmd() *cobra.Command {
 				out, close := getOut(cmd, sourcePath, outputName)
 				if out != nil {
 					defer close() //nolint errcheck
-					// path, _ := filepath.Abs(sourcePath)
 					// log.Debugf("Starting to process file %v", path)
 					config := configuration.NewConfiguration(
 						configuration.WithFilename(sourcePath),