--- a/make/go.mk
+++ b/make/go.mk
@@ -37,9 +37,20 @@ generate: install-pigeon
 	@if [ "pkg/parser/parser.go" -ot "pkg/parser/parser.peg" ]; then \
 		echo "generating the parser..."; \
 		pigeon ./pkg/parser/parser.peg > ./pkg/parser/parser.go; \
+		echo "done"; \
 	else \
 		echo "no need to regenerate the parser."; \
 	fi;
+
+.PHONY: verify-parser
+## verify that the parser was built with the latest version of pigeon, using the `optimize-grammar` option
+verify-parser: prebuild-checks clean
+ifneq ($(shell git diff --quiet pkg/parser/parser.go; echo $$?), 0)
+	@git diff pkg/parser/parser.go
+	$(error "pkg/parser/parser.go is uncommited or was generated with an older version of 'mna/pigeon' or without the '-optimize-parser' option enabled.")
+else
+	@echo "generated parser is ok"
+endif
 
 .PHONY: generate-optimized
 ## generate the .go file based on the asciidoc grammar
@@ -47,6 +58,7 @@ generate-optimized: install-pigeon
 	@if [ "pkg/parser/parser.go" -ot "pkg/parser/parser.peg" ]; then \
 		echo "generating the parser (optimized)..."; \
 		go generate ./...; \
+		echo "done"; \
 	else \
 		echo "no need to regenerate the parser."; \
 	fi;
@@ -68,33 +80,8 @@ build: prebuild-checks verify-parser generate-optimized
 
 PARSER_DIFF_STATUS :=
 
-.PHONY: verify-parser
-## verify that the parser was built with the latest version of pigeon, using the `optimize-grammar` option
-verify-parser: prebuild-checks
-ifneq ($(shell git diff --quiet pkg/parser/parser.go; echo $$?), 0)
-	@git diff pkg/parser/parser.go
-	$(error "parser was generated with an older version of 'mna/pigeon' or without the '-optimize-parser' option enabled.")
-else
-	@echo "generated parser is ok"
-endif
-
 .PHONY: install
 ## installs the binary executable in the $GOPATH/bin directory
 install: build
-	@cp $(BINARY_PATH) $(GOPATH)/bin
-
-.PHONY: quick-install
-## installs the binary executable in the $GOPATH/bin directory without prior tools setup and parser generation
-quick-install:
-	$(eval BUILD_COMMIT:=$(shell git rev-parse --short HEAD))
-	$(eval BUILD_TAG:=$(shell git tag --contains $(BUILD_COMMIT)))
-	$(eval BUILD_TIME:=$(shell date -u '+%Y-%m-%dT%H:%M:%SZ'))
-	@echo "building $(BINARY_PATH) (commit:$(BUILD_COMMIT) / tag:$(BUILD_TAG) / time:$(BUILD_TIME))"
-	@go build -ldflags \
-	  " -X github.com/bytesparadise/libasciidoc.BuildCommit=$(BUILD_COMMIT)\
-	    -X github.com/bytesparadise/libasciidoc.BuildTag=$(BUILD_TAG) \
-	    -X github.com/bytesparadise/libasciidoc.BuildTime=$(BUILD_TIME)" \
-	  -o $(BINARY_PATH) \
-	  cmd/libasciidoc/*.go
 	@cp $(BINARY_PATH) $(GOPATH)/bin