--- a/make/test.mk
+++ b/make/test.mk
@@ -10,7 +10,7 @@ install-gover:
 
 .PHONY: test
 ## run all tests excluding fixtures and vendored packages
-test: generate-optimized install-ginkgo
+test: clean generate-optimized install-ginkgo
 	@ginkgo -r --randomizeAllSpecs --randomizeSuites --failOnPending --trace --race --compilers=0
 
 COVERPKGS := $(shell go list ./... | grep -v vendor | paste -sd "," -)