--- a/pkg/parser/document_processing.go
+++ b/pkg/parser/document_processing.go
@@ -9,6 +9,8 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+const bufferSize = 10
+
 // ParseDocument parses the content of the reader identitied by the filename
 func ParseDocument(r io.Reader, config *configuration.Configuration, opts ...Option) (*types.Document, error) {
 	done := make(chan interface{})
@@ -37,8 +39,8 @@ func ParseDocument(r io.Reader, config *configuration.Configuration, opts ...Opt
 	if len(footnotes.Notes) > 0 {
 		doc.Footnotes = footnotes.Notes
 	}
-	if log.IsLevelEnabled(log.DebugLevel) {
-		log.Debugf("parsed document:\n%s", spew.Sdump(doc))
+	if log.IsLevelEnabled(log.InfoLevel) {
+		log.Infof("parsed document:\n%s", spew.Sdump(doc))
 	}
 	return doc, nil
 }