--- a/test/includes/attributes.adoc
+++ b/test/includes/attributes.adoc
@@ -1,0 +1,4 @@
+:author: Xavier
+:leveloffset: +1
+
+some content