… .github/workflows/ci-build.yml
… .golangci.yml
… cmd/libasciidoc/root_cmd_test.go
… go.sum
… libasciidoc_bench_test.go
… libasciidoc_test.go
… make/lint.mk
… pkg/parser/attribute_substitution_test.go
… pkg/parser/attributes_test.go
… pkg/parser/bench_test.go
… pkg/parser/blank_line_test.go
… pkg/parser/check_list_test.go
… pkg/parser/comment_test.go
… pkg/parser/cross_reference_test.go
… pkg/parser/delimited_block_admonition_test.go
… pkg/parser/delimited_block_example_test.go
… pkg/parser/delimited_block_fenced_test.go
… pkg/parser/delimited_block_listing_test.go
… pkg/parser/delimited_block_literal_test.go
… pkg/parser/delimited_block_markdown_quote_test.go
… pkg/parser/delimited_block_passthrough_test.go
… pkg/parser/delimited_block_quote_test.go
… pkg/parser/delimited_block_sidebar_test.go
… pkg/parser/delimited_block_source_test.go
… pkg/parser/delimited_block_verse_test.go
… pkg/parser/document_attributes_test.go
… pkg/parser/document_fragment_processing_test.go
… pkg/parser/document_header_test.go
… pkg/parser/document_processing__aggregate_test.go
… pkg/parser/document_processing__apply_substitutions.go