Fragments are blocks of contiguous lines separated by blank lines.
Blank lines within delimited blocks are preserved (hence they do not
split the fragments)
Once a fragment is read, it is sent on a channel to be processed
by the next stages of the pipeline until the aggregator which combines
all the fragments into the resulting `types.Document`, which
can then be rendered


BREAKING CHANGES:
- using pointers on all structs in `pkg/types`
- removing `types.Document.Attributes`, holding attributes in context
  as parsing/rendering progresses.
- `types.Paragraph.Lines` (`[][]interface{}`) is replaced by
  `types.Paragraph.Elements` (`[]interface{}`)
- all delimited blocks types are merged into `types.DelimitedBlock`
- all lists are merged into `types.List`
- removed the `Level` field in `types.ListElement` struct
- refactor document structure with header (level 0) and
  sections (level 1 to 5)

Fixes #843

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>