@xcoulon 2021-11-20
FWIW, running the benchs in [libasciidoc_bench_test.go](https://github.com/bytesparadise/libasciidoc/pull/855/files#diff-1f46b6c99207f5ba2d0ebab5bec483c934ee8da4966f89ee5cc758a3f286f305R30-R73) compared to the code tagged v0.6.0 gives the following improvements (via `benchstat`):

```
name                   old time/op    new time/op    delta
ParseBasicDocument-12    6.69ms ± 3%    0.68ms ± 3%  -89.79%  (p=0.000 n=9+9)
ParseLongDocument-12      295ms ± 2%      25ms ± 2%  -91.54%  (p=0.000 n=9+10)

name                   old alloc/op   new alloc/op   delta
ParseBasicDocument-12    2.01MB ± 0%    0.24MB ± 0%  -87.83%  (p=0.000 n=10+9)
ParseLongDocument-12     99.5MB ± 0%    11.5MB ± 0%  -88.44%  (p=0.000 n=9+10)

name                   old allocs/op  new allocs/op  delta
ParseBasicDocument-12     41.2k ± 0%      4.1k ± 0%  -89.97%  (p=0.001 n=8+9)
ParseLongDocument-12      2.06M ± 0%     0.20M ± 0%  -90.08%  (p=0.000 n=9+10)
```